refactor(utils): Move utils/get-moduules-metrics to modules/
