+++ packages/utils/src/modules/get-modules-metrics.js
