+++ packages/utils/src/modules/index.js
@@ -1,0 +1,1 @@
+export * from './get-modules-metrics';
